# docs/api-doc.yml
info:
  title: Salcedonie API
  version: 1.0.0
servers:
  - url: 'https://api.salcedonie.nakasar.me'
  - url: 'http://localhost:5000'
openapi: 3.0.0
...
paths:
...
  /events:
    get:
      operationId: getEvents
      tags:
        - events
      security:
        - user: []
        - application: []
      responses:
        '200':
          description: List of events the user may access.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventSummary'
    post:
      operationId: createEvent
      tags:
        - events
      security:
        - user: []
        - application: []
      requestBody:
        description: Event to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDetails'
  '/events/{eventId}':
    get:
      operationId: Get event details and last infos.
      tags:
        - events
      security:
        - user: []
        - application: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      responses:
        '200':
          description: Event requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDetails'
    patch:
      operationId: updateEvent
      tags:
        - events
      security:
        - user: []
        - application: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      requestBody:
        description: Event to update
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDetails'
    delete:
      operationId: archiveEvent
      tags:
        - events
      security:
        - user: []
        - application: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      responses:
        '204':
          description: Event archived
...
components:
  securitySchemes:
    user:
      description: User Bearer token
      type: http
      scheme: bearer
      bearerFormat: JWT
    application:
      description: Application API key
      type: apiKey
      name: API_KEY
      in: header
    admin:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EventSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        active:
          type: boolean
          description: Prevent user from posting actions if false.
        owner:
          type: string
          description: User that own the event.
      example:
        id: 5c1c06cd8d93224570fcc65b
        title: Prologue
        description: Le commencement de notre aventure !
        active: true
        owner: 5c1c06cd8d93224570fcc65b
    EventDetails:
      allOf:
        - $ref: '#/components/schemas/EventSummary'
        - type: object
          properties:
            text:
              type: string
          example:
            text: >-
              Voici un texte de description complet de l'évènement, avec
              éventuellement des intrusctions, etc.
    EventCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        text:
          type: string
        active:
          type: boolean
          description: Prevent user from posting actions if false.
        owner:
          type: string
          description: User that own sthe event. Default to bearer of the token.
      required:
        - title
        - description
      example:
        title: Prologue
        description: Le commencement de notre aventure !
        text: >-
          Voici un texte de description complet de l'évènement, avec
          éventuellement des intrusctions, etc.
        active: true
        owner: 5c1c06cd8d93224570fcc65b
...
  parameters:
...
    eventId:
      name: eventId
      in: path
      required: true
      description: Event ID
      schema:
        type: string
        example: 5c1c06cd8d93224570fcc65b
...